+++ src/main/java/az/gdg/mscomplaint/repository/ComplaintRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComplaintRepository extends JpaRepository<ComplaintEntity, Integer> {
+public interface ComplaintRepository extends JpaRepository<ComplaintEntity, Long> {
 
 }
