+++ src/main/java/az/gdg/mscomplaint/repository/ComplaintStatusRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComplaintStatusRepository extends CrudRepository<ComplaintStatusEntity, Integer> {
+public interface ComplaintStatusRepository extends CrudRepository<ComplaintStatusEntity, Long> {
 
 }
